2025_10_28_JGW
1. 장비 아이템 인벤토리 획득
2. UI_ObjPool -> Mgr_UI로 스크립트 옮김
